prompt/prompt runner/streamlit-prompt-runner/requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
jsonschema>=4.17.0
requests>=2.31.0
trimesh>=3.20.0
pymongo>=4.7.0
python-dotenv>=1.0.0
flask>=3.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
mongomock>=4.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
